requirements.txt
decorator==4.0.11
Django==1.8
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
nltk==3.2.4
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.2.0
pymongo==3.4.0
python-axolotl==0.1.39
python-axolotl-curve25519==0.1
python-dateutil==2.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.1
wcwidth==0.1.7
yowsup2==2.5.2
duckduckgo3
python-jenkins==0.4.14
